main: Add Vulkan support
For machines with dual-GPUs by 2 different vendors, setting the
VK_ICD_FILENAMES envvar reduces the GPUs it can detect to those of that
vendor, effectively forcing a particular GPU to be chosen.

For machines with dual-GPUs, the variable will have no effect, and
things will work as if no selection had been made, just like for NVidia
GPUs.

Closes: #23