main: Add Vulkan driver filtering
For machines with dual-GPUs by 2 different vendors, setting the
VK_LOADER_DRIVER_SELECT envvar reduces the GPUs it can detect to those
of that vendor, effectively forcing a particular GPU to be chosen.

For machines with dual-non-Mesa-GPUs, the variable will have no effect,
and things will work as if no selection had been made, just like for
NVidia GPUs.

Note that this will only work when the VulkanLoader in the application's
libraries are new enough to include this change:
https://github.com/KhronosGroup/Vulkan-Loader/pull/984

Closes: #23